created pullquote component in new file updating icon with svg
